Fixed Dead Defender Bug
When a defender is destroyed before damage is dealt, no attack should happen. This can be seen easiest when a player attacks another player's Hero while they have Misdirection -> Explosive Trap in play. Misdirection redirects to a minion with 2 or less health which is then destroyed by Explosive Trap. Since the defender is no longer alive, the attack never actually happens, so the attacker is free to attack a different target. This means that weapon charges and remaining attacks are never wasted on dead defenders.