Fixed Spells and Rarities forever
CardType and Rarity now contains an internal check for which type it is.
If extras are added (Some sort of super Mega-Legendary that is
equivalent to Legendary or whatever) then these can be updated for Card
Searches. Commons should accept Free, and Spells should accept Choose
Ones for everything now.